stop rendering autodependencies
cookbook autodependencies are bad to feed into depsolvers.  cookbooks
all automatically are dependencies of themselves, but some people still
do:

name 'foo'
depends 'foo'

what this can actually do is create depsolver loops where all the
versions of a cookbook create dependency loops which satsify this
constraint and make the depsolver work much harder than it should.

this patch just stops supermarket from publishing this information out
and starts to stop the pain.

additionally, there should be a migration added to clean these up and
there should be filtering on the input side.  knife and the chef server
also need to clean them up as well on the client and server side.

at the same time, though, we need to not completely break old cookbooks,
so we'll probably be stuck with maintaining cookbooks that have
auto dependencies in metadata.rb for some time, but we can gradually
make the tools all start ignoring those deps.