update css selector to be zebra-specific
As discussed in the comments on
https://github.com/chef/supermarket/pull/1045, the previous changeset
has unintended carrovers to other parts of the site using the same
template rendere fragment.

By adding a `.zebra` class, we can now define the type of simple_listing
at the `ul` class-level definition, and any other component can opt-in
to using this class.