Fix ICLA sigs to allow tests to inject a signed_at time
Without the check for signed_at having a value already, the tests'
setting of signed_at were being clobbered by the before_create
callback. The test's success relied upon the _ordering_ of the let
statements for the second let to be newer, not the specified date
given in the create statement.

Reordered ICLA lets to make the test fail, then added guard to the
callback to check whether signed_at already had a value.