Fix CCLA sigs to allow tests to inject a signed_at time
Without the check for signed_at having a value already, the tests'
setting of signed_at were being clobbered by the before_create
callback. The test's success relied upon the _ordering_ of the let
statements for the second let to be newer, not the specified date
given in the create statement.

Reordered CCLA & Organization lets to make the test fail, then:
* added guard to the callback to check whether signed_at already had
  a value
* organization lets needed to be told not to create the CCLAs defaulted
  in the organization factory - these were getting created with Time.now
  and dorking up the expectation of signatures only in the past
* renamed some variables within the spec to make roles in the scenario
  more clear