Add back the ability to override repo webhook callback URL
It's probably useful to keep this ability around, but not make it default.

Removed the callback URL from .env so that the default will be to use the
Rails url_helper for the callback.

Re-recorded the VCR cassettes for the subscriber spec which required
updating the repo org from 'gofullstack' to 'chef' so that the GitHub
API would be happier about being poked.

Updated VCR for CLA validation worker for 'chef' org.