Add a migration to ensure current DB extensions are enabled.
These extensions are already included in the schema. But pg_trgm was added
directly to schema.rb[1] leaving newly migrated databases (e.g. dev and
test) without the extensions being included in the resultant post-migration
schema.rb.

plpgsql is included from the PG adapter on platforms that need it, so it
is not included directly in this migration.

pg_trgrm is needed for PG search to work.

[1] https://github.com/chef/supermarket/commit/99d7990d85e36821c57d43d2861316f46c0407de