Specify that the badges_mask must be an integer
As part of the current interface between Badgeable and a badgeable thing,
not only must there be a `badges_mask` property on the thing, but it must
be an integer for the bit math to work.

Updated migration's date and to require the default value of 0.

Needed to include the badgeable_spec for cookbook_spec to run by itself.